refactor: move types closer to tool code
